写完querySingle和deleteMenu方法

+++ psi-java/psi-sysmanagement/src/main/java/com/zeroone/star/psisysmanagement/service/impl/SysMenuServiceImpl.java
@@ -13,6 +13,7 @@ import com.zeroone.star.project.vo.sysmanagement.menumanagement.MenuVO;
 import com.zeroone.star.psisysmanagement.entity.SysMenu;
 import com.zeroone.star.psisysmanagement.mapper.SysMenuMapper;
 import com.zeroone.star.psisysmanagement.service.ISysMenuService;
+import lombok.var;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -32,29 +33,17 @@ import java.util.List;
 public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> implements ISysMenuService {
 
     @Override
-    public JsonVO<MenuVO> querySingle(SingleMenuQuery singleMenuQuery) {
-        SysMenu sysMenu = baseMapper.selectById(singleMenuQuery.getId());
-        MenuVO menuVO = BeanUtil.copyProperties(sysMenu, MenuVO.class);
+    public JsonVO<List<MenuVO>> querySingle(SingleMenuQuery singleMenuQuery) {
 
-        return menuVO == null ?
-                JsonVO.create(null, ResultStatus.FAIL) : JsonVO.create(menuVO, ResultStatus.SUCCESS);
-    }
-
-    @Override
-    public JsonVO<List<MenuVO>> queryMenus(MenusQuery menusQuery) {
-
-        QueryWrapper<SysMenu> menuDTOQueryWrapper = new QueryWrapper<>();
-        menuDTOQueryWrapper.eq("parent_id", menusQuery.getParentId());
-
-        List<SysMenu> sysMenus = baseMapper.selectList(menuDTOQueryWrapper);
+        List<SysMenu> menuList = query().eq("parent_id", singleMenuQuery.getId()).list();
         ArrayList<MenuVO> menuVOS = new ArrayList<>();
-        for (SysMenu sysMenu : sysMenus) {
-            MenuVO menuVO = BeanUtil.copyProperties(sysMenu, MenuVO.class);
+        for (SysMenu sysMenu : menuList) {
+            MenuVO menuVO=BeanUtil.copyProperties(sysMenu,MenuVO.class);
             menuVOS.add(menuVO);
         }
-
         return JsonVO.create(menuVOS, ResultStatus.SUCCESS);
     }
+
 
     @Override
     @Transactional
@@ -62,72 +51,112 @@ public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> impl
 
         SysMenu sysMenu = BeanUtil.copyProperties(menuDTO, SysMenu.class);
         sysMenu.setCreateTime(LocalDateTime.now());
-
-        //获取父节点id
         String parentId = menuDTO.getParentId();
+
+        //判断是否在二级菜单下继续新增菜单
+        if (!parentId.equals("0")) {
+            double parentSort = query().eq("id", parentId).one().getSortNo();
+            String sortString = Double.toString(parentSort);
+            if (sortString.length() == 4 && !sortString.endsWith("0")) {
+                return JsonVO.fail(ResultStatus.FAIL);
+            }
+        }
+
+        setSortNo(sysMenu);
+
+        return save(sysMenu) ? JsonVO.success(ResultStatus.SUCCESS) : JsonVO.fail(ResultStatus.FAIL);
+
+    }
+
+    /**
+     * 根据menuDTO中的parentId，给sysMenu的sort_no赋值
+     *
+     * @param sysMenu
+     */
+    private void setSortNo(SysMenu sysMenu) {
+        //获取父节点id
+        String parentId = sysMenu.getParentId();
         if (parentId.equals("0")) {
 
-            double sort = 1.00;
+            int temp = 1;
+            double sort;
+            String sortString;
             //没有父节点时
-            while (true) {
+            while (true) {//修改小数点后第一位
+                sortString = temp + ".00";
+                //防止精度丢失，用String转为double
+                sort = Double.parseDouble(sortString);
                 //查询表中是否已有与sort等值的sort_no
-                QueryChainWrapper<SysMenu> menuEntity = query().eq("sort_no", sort);
-                if (menuEntity.isEmptyOfEntity()) {
-                    //如果没有就将sort值赋给新的menu的sort_no
-                    sysMenu.setSortNo(sort);
-                    break;
-                }
-                sort += 1.00;
+                if (checking(sysMenu, sort)) break;
+                temp++;
             }
 
         } else {
-
             //有父节点时，获取父节点的sort_no
             double parentSort = query().eq("id", parentId).one().getSortNo();
-            //将sort转换为字符串并去除小数点
-            String sortString = Double.toString(parentSort).replace(".", "");
+            //将sort转换为字符串
+            String sortString = Double.toString(parentSort);
+            //去除小数点
+            String replace = sortString.replace(".", "");
             //准备int数组sortArray
-            int[] sortArray = new int[sortString.length()];
+            int[] sortArray = new int[replace.length()];
             //将sortString每个数字传入数组
             for (int i = 0; i < sortArray.length; i++) {
-                sortArray[i] = sortString.charAt(i) - '0';
+                sortArray[i] = replace.charAt(i) - '0';
             }
 
             if (sortArray[1] == 0) {
-
                 //当父节点的Sort_no小数点后一位为0时，说明父节点为一级菜单，小数点第二位也必定为0，如1.00
-                double sort = parentSort + 0.10;
+                //temp为小数点后第一位
+                int temp = 1;
+                double sort;
+
                 while (true) {
+                    //修改小数点后第一位
+                    sortString = sortArray[0] + "." + temp + "0";
+                    //防止精度丢失，用String转为double
+                    sort = Double.parseDouble(sortString);
                     //尝试查询数据库中是否含有sort等值的sort_no
-                    QueryChainWrapper<SysMenu> menuEntity = query().eq("sort_no", sort);
-                    if (menuEntity.isEmptyOfEntity()) {
-                        //如果没有就将sort值赋给新的menu的sort_no
-                        sysMenu.setSortNo(sort);
-                        break;
-                    }
-                    sort += 0.10;
+                    if (checking(sysMenu, sort)) break;
+                    temp++;
                 }
             } else {
-
                 //如果小数点后第一位不为0，即父节点为二级菜单，如1.10；
                 //由于限制只能创建到三级菜单，所以父节点只有三种情况，为空，一级菜单，二级菜单
                 //故父节点sort_no小数点后第二位没有判断的必要
-                double sort = parentSort + 0.01;
+                int temp = 1;
+                double sort;
+
                 while (true) {
+                    //修改小数点后第二位
+                    sortString = sortArray[0] + "." + sortArray[1] + temp;
+                    //防止精度丢失，用String转为double
+                    sort = Double.parseDouble(sortString);
                     //尝试查询数据库中是否含有sort等值的sort_no
-                    QueryChainWrapper<SysMenu> menuEntity = query().eq("sort_no", sort);
-                    if (menuEntity.isEmptyOfEntity()) {
-                        //如果没有就将sort值赋给新的menu的sort_no
-                        sysMenu.setSortNo(sort);
-                        break;
-                    }
-                    sort += 0.01;
+                    if (checking(sysMenu, sort)) break;
+                    temp++;
                 }
             }
         }
 
-        return save(sysMenu) ? JsonVO.success(ResultStatus.SUCCESS) : JsonVO.fail(ResultStatus.FAIL);
-
+    }
+    /**
+     * setSortNo方法中判断数据库中是否有重复sort_no
+     *
+     * @param sysMenu
+     * @param sort
+     * @return
+     */
+    private boolean checking(SysMenu sysMenu, double sort) {
+        QueryWrapper<SysMenu> queryWrapper = new QueryWrapper<>();
+        queryWrapper.select("sort_no").eq("sort_no", sort);
+        baseMapper.selectOne(queryWrapper);
+        if (baseMapper.selectOne(queryWrapper) == null) {
+            //如果没有就将sort值赋给新的menu的sort_no
+            sysMenu.setSortNo(sort);
+            return true;
+        }
+        return false;
     }
 
     @Transactional
@@ -146,10 +175,20 @@ public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> impl
     @Override
     public JsonVO<ResultStatus> deleteMenu(SingleMenuQuery singleMenuQuery) {
 
-        SysMenu sysMenu = BeanUtil.copyProperties(singleMenuQuery, SysMenu.class);
 
-        int num = baseMapper.deleteById(sysMenu);
 
-        return num >= 1 ? JsonVO.success(ResultStatus.SUCCESS) : JsonVO.fail(ResultStatus.FAIL);
+        List<SysMenu> listSysMenu= query().eq("parent_id", singleMenuQuery.getId()).list();
+        if (listSysMenu.isEmpty()){
+            SysMenu sysMenu = BeanUtil.copyProperties(singleMenuQuery, SysMenu.class);
+            int num = baseMapper.deleteById(sysMenu);
+
+            return num >= 1 ? JsonVO.success(ResultStatus.SUCCESS) : JsonVO.fail(ResultStatus.FAIL);
+        }
+        baseMapper.deleteById(singleMenuQuery.getId());
+
+        boolean result = removeByIds(listSysMenu);
+        return result==true? JsonVO.success(ResultStatus.SUCCESS) : JsonVO.fail(ResultStatus.FAIL);
+
     }
+
 }
